Rename Loader to Login.
